+++ src/Resources/contao/dca/tl_synapsis_like.php
@@ -1,0 +1,58 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Dieses Bundle stellt das Forum "Synapsis" fuer Contao 4.13 und Contao 5 bereit.
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+use Contao\DC_Table;
+use Schachbulle\ContaoSynapsisBundle\SchachbulleContaoSynapsisBundle;
+
+/*
+ * Tabelle tl_synapsis_like
+ *
+ * Verknuepfungstabelle: welches Mitglied welchen Beitrag mit "Gefaellt mir"
+ * markiert hat. Reine Frontend-Pflege; die DCA dient allein der
+ * Tabellenerzeugung durch contao:migrate.
+ */
+$GLOBALS['TL_DCA']['tl_synapsis_like'] = array
+(
+    'config' => array
+    (
+        'dataContainer' => SchachbulleContaoSynapsisBundle::isContao5() ? DC_Table::class : 'Table',
+        'notEditable'   => true,
+        'closed'        => true,
+        'sql' => array
+        (
+            'keys' => array
+            (
+                'id'          => 'primary',
+                'member,post' => 'unique',
+                'post'        => 'index',
+            ),
+        ),
+    ),
+
+    'fields' => array
+    (
+        'id' => array
+        (
+            'sql' => "int(10) unsigned NOT NULL auto_increment",
+        ),
+        'tstamp' => array
+        (
+            'sql' => "int(10) unsigned NOT NULL default 0",
+        ),
+        'member' => array
+        (
+            'sql' => "int(10) unsigned NOT NULL default 0",
+        ),
+        'post' => array
+        (
+            'sql' => "int(10) unsigned NOT NULL default 0",
+        ),
+    ),
+);